+++ src/css/base/common.css
@@ -67,13 +67,13 @@ html {
 
 @media screen and (min-width: 375px) and (max-width: 767px) {
   .container {
-    width: 375px;
+    min-width: 375px;
   }
 }
 
 @media screen and (min-width: 768px) {
   .container {
-    width: 768px;
+    min-width: 768px;
     padding-left: 32px;
     padding-right: 32px;
   }
@@ -81,6 +81,6 @@ html {
 
 @media screen and (min-width: 1280px) {
   .container {
-    width: 1280px;
+    min-width: 1280px;
   }
 }
