+++ src/css/main.css
@@ -7,4 +7,5 @@
 @import url('./layout/favorites.css');
 @import url('./layout/footer.css');
 @import url('./layout/header.css');
+@import url('./layout/burger-modal.css');
 @import url('./layout/hero.css');
