+++ src/main.js
@@ -1,2 +1,3 @@
-import{googleSignin, logOut} from "./js/header.js"
-export {googleSignin, logOut};
+import{googleSignin, logOut} from "./js/authentication.js"
+import {openMenu, closeMenu} from "./js/burger"
+export {googleSignin, logOut, openMenu, closeMenu};
