+++ src/partials/burger.html
@@ -1,30 +1,34 @@
-<did class="burger-modal">
-    <svg  xmlns="http://www.w3.org/2000/svg" width="32px" height="32"> 
-        <use href="./img/icons.svg#icon-close" fill="white" />
-      </svg>
-      <ul>
-        <li>Home</li>
-        <li>favorites</li>
+<div class="burger-modal visually-hidden" id="burger">
+   <button class="burger-button-close" id ="burger-button-close">
+       <svg  xmlns="http://www.w3.org/2000/svg" width="16px" height="16"> 
+           <use href="./img/icons.svg#icon-close"  />
+         </svg>
+   </button>
+      <ul class="burger-pages-list">
+        <li class="burger-pages-list-item">Home</li>
+        <li class="burger-pages-list-item">favorites</li>
       </ul>
-      <ul class ="social-links">
-        <li>
-    
-          <svg  xmlns="http://www.w3.org/2000/svg" width="32px" height="32"> 
-            <use href="./img/icons.svg#icon-facebook" fill="#242424" />
-          </svg>
+      <ul class ="social-links-list">
+          <li >
+          <a class="social-links-list-item" href="https://www.facebook.com/goITclub/" target="_blank">
+            <svg xmlns="http://www.w3.org/2000/svg" width="32px" height="32"> 
+                <use href="./img/icons.svg#icon-facebook"  fill="white"/>
+              </svg>
+            </a>
         </li >
-        <li>
-    
+        <li >
+        <a  class="social-links-list-item" href="https://www.instagram.com/goitclub/" target="_blank">
           <svg  xmlns="http://www.w3.org/2000/svg" width="32px" height="32"> 
-            <use href="./img/icons.svg#icon-instagram" fill="#242424" />
+            <use href="./img/icons.svg#icon-instagram"  />
           </svg>
-        </li>
-        <li>
-    
+        </a>
+    </li>
+    <li >
+        <a class="social-links-list-item" href="https://www.youtube.com/c/GoIT" target="_blank">
           <svg  xmlns="http://www.w3.org/2000/svg" width="32px" height="32"> 
-            <use href="./img/icons.svg#icon-youtube" fill="#242424" />
+            <use href="./img/icons.svg#icon-youtube" />
           </svg>
-        </li>
-       
+        </a>
+    </li>
       </ul>
-</did>
+</div>
