+++ src/partials/header.html
@@ -1,36 +1,42 @@
-<header>
-  <img src = "./img/your-energy.svg" alt="logo"/>
-  <ul>
-    <li>Home</li>
-    <li>favorites</li>
-  </ul>
-  <p id="user"></p>
-  <button  id="sign-in">Sign in</button>
-  <button  id="sign-out" class="display-none">Sign out</button>
-  <ul class ="social-links">
-    <li>
+<header class="container">
+  <div class="header">
 
+    <img src = "./img/your-energy.svg" alt="logo"/>
+    <ul class="page-navigattion">
+      <li class="page-navigattion-item active" id="home"><span class="black" id="home-text">Home</span></li>
+      <li class="page-navigattion-item" id="favorites"><span  id="favorites-text">Favorites</span></li>
+    </ul>
+    <p id="user"></p>
+    <button id="sign-in">
+      <img src = "./img/icon-google.svg" alt="google" width="24px" height="24"/>
+      Sign in</button>
+  
+    <button  id="sign-out" class="display-none">Sign out</button>
+    <ul class ="social-links">
+      <li>
+  
+        <svg  xmlns="http://www.w3.org/2000/svg" width="32px" height="32"> 
+          <use href="./img/icons.svg#icon-facebook" fill="#242424" />
+        </svg>
+      </li >
+      <li>
+  
+        <svg  xmlns="http://www.w3.org/2000/svg" width="32px" height="32"> 
+          <use href="./img/icons.svg#icon-instagram" fill="#242424" />
+        </svg>
+      </li>
+      <li>
+  
+        <svg  xmlns="http://www.w3.org/2000/svg" width="32px" height="32"> 
+          <use href="./img/icons.svg#icon-youtube" fill="#242424" />
+        </svg>
+      </li>
+    </ul>
+    <button class="burger-button" id="burger-button">
       <svg  xmlns="http://www.w3.org/2000/svg" width="32px" height="32"> 
-        <use href="./img/icons.svg#icon-facebook" fill="#242424" />
+        <use href="./img/icons.svg#icon-burger-menu" stroke="#242424" />
       </svg>
-    </li >
-    <li>
-
-      <svg  xmlns="http://www.w3.org/2000/svg" width="32px" height="32"> 
-        <use href="./img/icons.svg#icon-instagram" fill="#242424" />
-      </svg>
-    </li>
-    <li>
-
-      <svg  xmlns="http://www.w3.org/2000/svg" width="32px" height="32"> 
-        <use href="./img/icons.svg#icon-youtube" fill="#242424" />
-      </svg>
-    </li>
-  </ul>
-  <button class="display-none">
-    <svg  xmlns="http://www.w3.org/2000/svg" width="32px" height="32"> 
-      <use href="./img/icons.svg#icon-burger-menu" fill="#242424" />
-    </svg>
-  </button>
+    </button>
+  </div>
 </header>
